feat(unstable): worked on class implementations

+++ MemberSystem/src/main/java/com/isacariotsystems/MemberSystem/entity/Rank.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Rank {
     @Id
-    private Long rankId;
+    private Integer rankId;
 
     private String description;
